Feature ETP-3546: Refresh PO pipeline and move context switcher
Re-ran pipeline for Purchase Order (203 fields, 403 tests pass).
Moved org/role context switcher to "Your company" in sidebar header.
