+++ artifacts/purchase-order/schema-raw.json
@@ -3752,6 +3752,6 @@
   "meta": {
     "version": "0.1.0",
     "checksum": "cd96bd6b",
-    "extractedAt": "2026-03-13T19:41:02.425Z"
+    "extractedAt": "2026-03-14T05:39:50.772Z"
   }
 }
